LayoutTests/imported/w3c/web-platform-tests/fetch/api/request/request-bad-port.any.worker-expected.txt

CONSOLE MESSAGE: Not allowed to use restricted network port 1: http://example.com:1/
CONSOLE MESSAGE: Not allowed to use restricted network port 7: http://example.com:7/
CONSOLE MESSAGE: Not allowed to use restricted network port 9: http://example.com:9/
CONSOLE MESSAGE: Not allowed to use restricted network port 11: http://example.com:11/
CONSOLE MESSAGE: Not allowed to use restricted network port 13: http://example.com:13/
CONSOLE MESSAGE: Not allowed to use restricted network port 15: http://example.com:15/
CONSOLE MESSAGE: Not allowed to use restricted network port 17: http://example.com:17/
CONSOLE MESSAGE: Not allowed to use restricted network port 19: http://example.com:19/
CONSOLE MESSAGE: Not allowed to use restricted network port 20: http://example.com:20/
CONSOLE MESSAGE: Not allowed to use restricted network port 21: http://example.com:21/
CONSOLE MESSAGE: Not allowed to use restricted network port 22: http://example.com:22/
CONSOLE MESSAGE: Not allowed to use restricted network port 23: http://example.com:23/
CONSOLE MESSAGE: Not allowed to use restricted network port 25: http://example.com:25/
CONSOLE MESSAGE: Not allowed to use restricted network port 37: http://example.com:37/
CONSOLE MESSAGE: Not allowed to use restricted network port 42: http://example.com:42/
CONSOLE MESSAGE: Not allowed to use restricted network port 43: http://example.com:43/
CONSOLE MESSAGE: Not allowed to use restricted network port 53: http://example.com:53/
CONSOLE MESSAGE: Not allowed to use restricted network port 69: http://example.com:69/
CONSOLE MESSAGE: Not allowed to use restricted network port 77: http://example.com:77/
CONSOLE MESSAGE: Not allowed to use restricted network port 79: http://example.com:79/
CONSOLE MESSAGE: Not allowed to use restricted network port 87: http://example.com:87/
CONSOLE MESSAGE: Not allowed to use restricted network port 95: http://example.com:95/
CONSOLE MESSAGE: Not allowed to use restricted network port 101: http://example.com:101/
CONSOLE MESSAGE: Not allowed to use restricted network port 102: http://example.com:102/
CONSOLE MESSAGE: Not allowed to use restricted network port 103: http://example.com:103/
CONSOLE MESSAGE: Not allowed to use restricted network port 104: http://example.com:104/
CONSOLE MESSAGE: Not allowed to use restricted network port 109: http://example.com:109/
CONSOLE MESSAGE: Not allowed to use restricted network port 110: http://example.com:110/
CONSOLE MESSAGE: Not allowed to use restricted network port 111: http://example.com:111/
CONSOLE MESSAGE: Not allowed to use restricted network port 113: http://example.com:113/
CONSOLE MESSAGE: Not allowed to use restricted network port 115: http://example.com:115/
CONSOLE MESSAGE: Not allowed to use restricted network port 117: http://example.com:117/
CONSOLE MESSAGE: Not allowed to use restricted network port 119: http://example.com:119/
CONSOLE MESSAGE: Not allowed to use restricted network port 123: http://example.com:123/
CONSOLE MESSAGE: Not allowed to use restricted network port 135: http://example.com:135/
CONSOLE MESSAGE: Not allowed to use restricted network port 137: http://example.com:137/
CONSOLE MESSAGE: Not allowed to use restricted network port 139: http://example.com:139/
CONSOLE MESSAGE: Not allowed to use restricted network port 143: http://example.com:143/
CONSOLE MESSAGE: Not allowed to use restricted network port 161: http://example.com:161/
CONSOLE MESSAGE: Not allowed to use restricted network port 179: http://example.com:179/
CONSOLE MESSAGE: Not allowed to use restricted network port 389: http://example.com:389/
CONSOLE MESSAGE: Not allowed to use restricted network port 427: http://example.com:427/
CONSOLE MESSAGE: Not allowed to use restricted network port 465: http://example.com:465/
CONSOLE MESSAGE: Not allowed to use restricted network port 512: http://example.com:512/
CONSOLE MESSAGE: Not allowed to use restricted network port 513: http://example.com:513/
CONSOLE MESSAGE: Not allowed to use restricted network port 514: http://example.com:514/
CONSOLE MESSAGE: Not allowed to use restricted network port 515: http://example.com:515/
CONSOLE MESSAGE: Not allowed to use restricted network port 526: http://example.com:526/
CONSOLE MESSAGE: Not allowed to use restricted network port 530: http://example.com:530/
CONSOLE MESSAGE: Not allowed to use restricted network port 531: http://example.com:531/
CONSOLE MESSAGE: Not allowed to use restricted network port 532: http://example.com:532/
CONSOLE MESSAGE: Not allowed to use restricted network port 540: http://example.com:540/
CONSOLE MESSAGE: Not allowed to use restricted network port 548: http://example.com:548/
CONSOLE MESSAGE: Not allowed to use restricted network port 554: http://example.com:554/
CONSOLE MESSAGE: Not allowed to use restricted network port 556: http://example.com:556/
CONSOLE MESSAGE: Not allowed to use restricted network port 563: http://example.com:563/
CONSOLE MESSAGE: Not allowed to use restricted network port 587: http://example.com:587/
CONSOLE MESSAGE: Not allowed to use restricted network port 601: http://example.com:601/
CONSOLE MESSAGE: Not allowed to use restricted network port 636: http://example.com:636/
CONSOLE MESSAGE: Not allowed to use restricted network port 989: http://example.com:989/
CONSOLE MESSAGE: Not allowed to use restricted network port 990: http://example.com:990/
CONSOLE MESSAGE: Not allowed to use restricted network port 993: http://example.com:993/
CONSOLE MESSAGE: Not allowed to use restricted network port 995: http://example.com:995/
CONSOLE MESSAGE: Not allowed to use restricted network port 1719: http://example.com:1719/
CONSOLE MESSAGE: Not allowed to use restricted network port 1720: http://example.com:1720/
CONSOLE MESSAGE: Not allowed to use restricted network port 1723: http://example.com:1723/
CONSOLE MESSAGE: Not allowed to use restricted network port 2049: http://example.com:2049/
CONSOLE MESSAGE: Not allowed to use restricted network port 3659: http://example.com:3659/
CONSOLE MESSAGE: Not allowed to use restricted network port 4045: http://example.com:4045/
CONSOLE MESSAGE: Not allowed to use restricted network port 5060: http://example.com:5060/
CONSOLE MESSAGE: Not allowed to use restricted network port 5061: http://example.com:5061/
CONSOLE MESSAGE: Not allowed to use restricted network port 6000: http://example.com:6000/
CONSOLE MESSAGE: Not allowed to use restricted network port 6566: http://example.com:6566/
CONSOLE MESSAGE: Not allowed to use restricted network port 6665: http://example.com:6665/
CONSOLE MESSAGE: Not allowed to use restricted network port 6666: http://example.com:6666/
CONSOLE MESSAGE: Not allowed to use restricted network port 6667: http://example.com:6667/
CONSOLE MESSAGE: Not allowed to use restricted network port 6668: http://example.com:6668/
CONSOLE MESSAGE: Not allowed to use restricted network port 6669: http://example.com:6669/
CONSOLE MESSAGE: Not allowed to use restricted network port 6697: http://example.com:6697/
CONSOLE MESSAGE: Not allowed to use restricted network port 10080: http://example.com:10080/

PASS Request on bad port 1 should throw TypeError.
PASS Request on bad port 7 should throw TypeError.
PASS Request on bad port 9 should throw TypeError.
PASS Request on bad port 11 should throw TypeError.
PASS Request on bad port 13 should throw TypeError.
PASS Request on bad port 15 should throw TypeError.
PASS Request on bad port 17 should throw TypeError.
PASS Request on bad port 19 should throw TypeError.
PASS Request on bad port 20 should throw TypeError.
PASS Request on bad port 21 should throw TypeError.
PASS Request on bad port 22 should throw TypeError.
PASS Request on bad port 23 should throw TypeError.
PASS Request on bad port 25 should throw TypeError.
PASS Request on bad port 37 should throw TypeError.
PASS Request on bad port 42 should throw TypeError.
PASS Request on bad port 43 should throw TypeError.
PASS Request on bad port 53 should throw TypeError.
PASS Request on bad port 69 should throw TypeError.
PASS Request on bad port 77 should throw TypeError.
PASS Request on bad port 79 should throw TypeError.
PASS Request on bad port 87 should throw TypeError.
PASS Request on bad port 95 should throw TypeError.
PASS Request on bad port 101 should throw TypeError.
PASS Request on bad port 102 should throw TypeError.
PASS Request on bad port 103 should throw TypeError.
PASS Request on bad port 104 should throw TypeError.
PASS Request on bad port 109 should throw TypeError.
PASS Request on bad port 110 should throw TypeError.
PASS Request on bad port 111 should throw TypeError.
PASS Request on bad port 113 should throw TypeError.
PASS Request on bad port 115 should throw TypeError.
PASS Request on bad port 117 should throw TypeError.
PASS Request on bad port 119 should throw TypeError.
PASS Request on bad port 123 should throw TypeError.
PASS Request on bad port 135 should throw TypeError.
PASS Request on bad port 137 should throw TypeError.
PASS Request on bad port 139 should throw TypeError.
PASS Request on bad port 143 should throw TypeError.
PASS Request on bad port 161 should throw TypeError.
PASS Request on bad port 179 should throw TypeError.
PASS Request on bad port 389 should throw TypeError.
PASS Request on bad port 427 should throw TypeError.
PASS Request on bad port 465 should throw TypeError.
PASS Request on bad port 512 should throw TypeError.
PASS Request on bad port 513 should throw TypeError.
PASS Request on bad port 514 should throw TypeError.
PASS Request on bad port 515 should throw TypeError.
PASS Request on bad port 526 should throw TypeError.
PASS Request on bad port 530 should throw TypeError.
PASS Request on bad port 531 should throw TypeError.
PASS Request on bad port 532 should throw TypeError.
PASS Request on bad port 540 should throw TypeError.
PASS Request on bad port 548 should throw TypeError.
PASS Request on bad port 554 should throw TypeError.
PASS Request on bad port 556 should throw TypeError.
PASS Request on bad port 563 should throw TypeError.
PASS Request on bad port 587 should throw TypeError.
PASS Request on bad port 601 should throw TypeError.
PASS Request on bad port 636 should throw TypeError.
PASS Request on bad port 989 should throw TypeError.
PASS Request on bad port 990 should throw TypeError.
PASS Request on bad port 993 should throw TypeError.
PASS Request on bad port 995 should throw TypeError.
PASS Request on bad port 1719 should throw TypeError.
PASS Request on bad port 1720 should throw TypeError.
PASS Request on bad port 1723 should throw TypeError.
PASS Request on bad port 2049 should throw TypeError.
PASS Request on bad port 3659 should throw TypeError.
PASS Request on bad port 4045 should throw TypeError.
PASS Request on bad port 5060 should throw TypeError.
PASS Request on bad port 5061 should throw TypeError.
PASS Request on bad port 6000 should throw TypeError.
PASS Request on bad port 6566 should throw TypeError.
PASS Request on bad port 6665 should throw TypeError.
PASS Request on bad port 6666 should throw TypeError.
PASS Request on bad port 6667 should throw TypeError.
PASS Request on bad port 6668 should throw TypeError.
PASS Request on bad port 6669 should throw TypeError.
PASS Request on bad port 6697 should throw TypeError.
PASS Request on bad port 10080 should throw TypeError.
